LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/report-blocked-uri-cross-origin.sub-expected.txt

CONSOLE MESSAGE: [Report Only] Refused to load http://www1.localhost:8800/content-security-policy/support/pass.png because it does not appear in the img-src directive of the Content Security Policy.
Blocked access to external URL http://www1.localhost:8800/content-security-policy/support/pass.png
CONSOLE MESSAGE: [Report Only] Refused to load http://www1.localhost:8800/content-security-policy/support/pass.png because it does not appear in the img-src directive of the Content Security Policy.
Blocked access to external URL http://www1.localhost:8800/content-security-policy/support/pass.png


PASS Violation report status OK.
